Exception Handling

+++ src/main/java/tech/biblioteca/Biblioteca/controller/AuthorController.java
@@ -16,6 +16,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import jakarta.validation.Valid;
 import tech.biblioteca.Biblioteca.entities.Author;
+import tech.biblioteca.Biblioteca.exceptions.ObjectNotFound;
 import tech.biblioteca.Biblioteca.services.AuthorService;
 
 
@@ -45,12 +46,18 @@ return authorService.getAll();
 @GetMapping("/{id}")
 public ResponseEntity<Author> getByid(@PathVariable Long id){
 Author obj = authorService.getById(id);
+if(obj == null) {
+	throw new ObjectNotFound("User not find with the id: " + id);
+}
 return ResponseEntity.ok().body(obj);
 }
 
 @DeleteMapping("/{id}")
 public List<Author> delete(@PathVariable Long id){
 List<Author> obj = authorService.delete(id);
+if(obj == null) {
+	throw new ObjectNotFound("User not find with the id: " + id);
+}
 return obj;
 }
